Finally connected with facebook, changed drawing and window calling
I added my first round of Facebook integration. For now you'll need to
get your access token from the Faceook developers page (a handy link is
included on the login page). I created a 'Switch' class to better call
windows and allow for you to return to a previous window. Next up is
the the photo upload!